Boon fixes, part 3: fix for #48
This change moves the impenetrable boon fix over to the cbfm_update_initiatives function with its friend, the slayer of champions boon fix. Both fixes have been updated to apply to reinforcing generals as well as primary generals. Tested in 2.1.1, both work. Note that I left the old impenetrable boon fix in the main initiative unlock script so as not to unnecessarily change that file, but that fix is being overwritten by this one at game load now.