Fix for #120 (tested)
I went through all of the blessed spawning units but only found two omissions: Missile resistance for the Blessed Stegadon and Primal Instincts for Blessed Horned Ones. In the latter case, I believe it used to be a benefit to be missing Primal Instincts, because that ability used to come with Rampage. Now that Primal Instincts is purely positive, it makes sense to reinstate it, but it's worth noting that Blessed Horned Ones now possess no other benefit beyond the standard extra HP.