Fix for #176 (severity only, reward not a bug)
Tested in 2.1.1, this will set Belegar's primary grudge to 10 severity, which is how it was in ME and almost certainly where it should be. I had to go back and check, but Belegar actually never did receive an added reward for completing this grudge in WH2: K8P is apparently reward enough.

One final note: Belegar used to have a fourth default grudge to take Ekrund, but it looks like this one was intentionally removed and turned into an easier random grudge, presumably to limit early-game penalties for a faction that already has a notoriously difficult start.