Fix for issue #265
Added overcasted spells to effects that reduce costs or cooldowns by a certain amount. This affects Be'lakor, Kairos (Half the shards in his skill tree worked since they were linked to older effects which already had the overcast spells which made it obvious this is an oversight), Vilitch and his tech tree, Festus and his tech tree and ascended sorcerer traits.