Expanded functionality for all Chaos
It turns out this problem was not only constrained to the Chaos DLC lords' military-focused techs. Instead, every Chaos tech that works via additional effects which are multiplied by active Chaos gifts does not persist when loading a game (or returning to campaign from a battle). This script and supporting DB content has been expanded to support redoing the functionality for all of these techs, allowing their states to be properly saved and loaded.