Lua duct tape for #26
This is really two fixes in one: For the AI, it checks DE generals every turn for extra name of power traits and removes them if found (it uses saved values to keep track of the original name so it is preserved). For the player, it adds a dilemma listener that fires whenever there is a name of power question, and if the general being offered already has a name of power, it quickly closes the dilemma by forcing the first choice, removing the trait associated with that choice, then clearing all UI elements referencing the whole transaction, leaving the player back where they started before the bug manifested.