New fix for #1390
This fix is based on the current version of CA's script rather than the old one. The only change is copying the ancillary table before randomizing, which preserves the original between drops. Unfortunately, there (still) is no way to fix this without entirely rewriting this script on account of the local variables.