Band-aid for #1495
This fix changes the two Cacophonous Assault (the regular and upgraded version) army abilities to be active because they will not properly trigger if passive. To maintain intended behavior of the ability automatically firing at battle start, a battle script is included that fires it immediately for all Beastmen factions in a battle that have it, making it effectively passive even if it technically isn't. The upgraded version of the ability was also assigned the wrong target type; this has been fixed as well.